refactor(api): move notification-settings to modules/ as controller, model, service
