+++ apps/api/src/notifications/send.ts
@@ -1,6 +1,6 @@
 import { sendEmail, emailBody, type EmailConfig, type SendResult } from '@repo/mailer';
 import { getProjectEmailConfig } from '@repo/auth';
-import { emailSource, type NotificationConfig } from '../notification-settings/store';
+import { emailSource, type NotificationConfig } from '#modules/notification-settings/service';
 import type { DeliveryPayload } from './outbound';
 import { getInstanceBotConfig, isInstanceBotUsable } from '../telegram/store';
