+++ apps/api/src/planner.ts
@@ -32,7 +32,7 @@ import { agentScheduleRoutes } from './modules/agents/schedules';
 import { agentRunnerRoutes } from './modules/agents/runner';
 import { agentChatRoutes } from './modules/agents/chat';
 import { notificationRoutes } from './notifications/routes';
-import { notificationSettingsRoutes } from './notification-settings/routes';
+import { notificationSettingsRoutes } from './modules/notification-settings';
 import { notificationPreferenceRoutes } from './modules/notification-preferences';
 import { userPreferenceRoutes } from './modules/user-preferences';
 import { telegramRoutes } from './telegram/routes';
